Refactor and remove validation message handler
