+++ src/main/java/com/networknt/schema/BaseJsonValidator.java
@@ -17,61 +17,65 @@
 package com.networknt.schema;
 
 import com.fasterxml.jackson.databind.JsonNode;
-import com.fasterxml.jackson.databind.node.ObjectNode;
 import com.networknt.schema.annotation.JsonNodeAnnotation;
-import com.networknt.schema.i18n.DefaultMessageSource;
-import com.networknt.schema.i18n.MessageSource;
 
 import org.slf4j.Logger;
 
 import java.util.Collection;
-import java.util.Iterator;
 import java.util.Map;
-import java.util.Set;
 import java.util.function.Consumer;
 
 /**
  * Base {@link JsonValidator}. 
  */
-public abstract class BaseJsonValidator extends ValidationMessageHandler implements JsonValidator {
-    protected final boolean suppressSubSchemaRetrieval;
-
+public abstract class BaseJsonValidator implements JsonValidator {
     protected final JsonNode schemaNode;
 
     protected final ValidationContext validationContext;
 
+    protected final Keyword keyword;
+    protected final JsonSchema parentSchema;
+    protected final SchemaLocation schemaLocation;
+    protected final Map<String, String> errorMessage;
+
+    protected final JsonNodePath evaluationPath;
+    protected final JsonSchema evaluationParentSchema;
+
+    // Pending removal
+    protected final ErrorMessageType errorMessageType;
+
     public BaseJsonValidator(SchemaLocation schemaLocation, JsonNodePath evaluationPath, JsonNode schemaNode,
             JsonSchema parentSchema, ValidatorTypeCode validatorType, ValidationContext validationContext) {
-        this(schemaLocation, evaluationPath, schemaNode, parentSchema, validatorType, validatorType, validationContext,
-                false);
+        this(schemaLocation, evaluationPath, schemaNode, parentSchema, validatorType, validatorType, validationContext);
     }
 
     public BaseJsonValidator(SchemaLocation schemaLocation, JsonNodePath evaluationPath, JsonNode schemaNode,
             JsonSchema parentSchema, ErrorMessageType errorMessageType, Keyword keyword,
-            ValidationContext validationContext, boolean suppressSubSchemaRetrieval) {
-        super(errorMessageType,
-                (validationContext != null && validationContext.getConfig() != null)
-                        ? validationContext.getConfig().getErrorMessageKeyword()
-                        : null,
-                (validationContext != null && validationContext.getConfig() != null)
-                        ? validationContext.getConfig().getMessageSource()
-                        : DefaultMessageSource.getInstance(),
-                keyword,
-                parentSchema, schemaLocation, evaluationPath);
+            ValidationContext validationContext) {
         this.validationContext = validationContext;
         this.schemaNode = schemaNode;
-        this.suppressSubSchemaRetrieval = suppressSubSchemaRetrieval;
+
+        this.keyword = keyword;
+        this.parentSchema = parentSchema;
+        this.schemaLocation = schemaLocation;
+        if (keyword != null && parentSchema != null && validationContext.getConfig().getErrorMessageKeyword() != null) {
+            this.errorMessage = ErrorMessages.getErrorMessage(parentSchema,
+                    validationContext.getConfig().getErrorMessageKeyword(), keyword.getValue());
+        } else {
+            this.errorMessage = null;
+        }
+
+        this.errorMessageType = errorMessageType;
+        this.evaluationPath = evaluationPath;
+        this.evaluationParentSchema = null;
     }
 
     /**
      * Constructor to create a copy using fields.
      *
-     * @param suppressSubSchemaRetrieval to suppress sub schema retrieval
      * @param schemaNode the schema node
      * @param validationContext the validation context
      * @param errorMessageType the error message type
-     * @param errorMessageKeyword the error message keyword
-     * @param messageSource the message source
      * @param keyword the keyword
      * @param parentSchema the parent schema
      * @param schemaLocation the schema location
@@ -81,45 +85,27 @@ public abstract class BaseJsonValidator extends ValidationMessageHandler impleme
      */
     protected BaseJsonValidator(
             /* Below from BaseJsonValidator */
-            boolean suppressSubSchemaRetrieval,
             JsonNode schemaNode,
             ValidationContext validationContext,
             /* Below from ValidationMessageHandler */
             ErrorMessageType errorMessageType,
-            String errorMessageKeyword,
-            MessageSource messageSource,
             Keyword keyword,
             JsonSchema parentSchema,
             SchemaLocation schemaLocation,
             JsonNodePath evaluationPath,
             JsonSchema evaluationParentSchema,
             Map<String, String> errorMessage) {
-        super(errorMessageType, errorMessageKeyword, messageSource, keyword,
-                parentSchema, schemaLocation, evaluationPath, evaluationParentSchema, errorMessage);
-        this.suppressSubSchemaRetrieval = suppressSubSchemaRetrieval;
         this.schemaNode = schemaNode;
         this.validationContext = validationContext;
-    }
+        
+        this.keyword = keyword;
+        this.parentSchema = parentSchema;
+        this.schemaLocation = schemaLocation;
+        this.errorMessage = errorMessage;
 
-    private static JsonSchema obtainSubSchemaNode(final JsonNode schemaNode, final ValidationContext validationContext) {
-        final JsonNode node = schemaNode.get("id");
-
-        if (node == null) {
-            return null;
-        }
-
-        if (node.equals(schemaNode.get("$schema"))) {
-            return null;
-        }
-
-        final String text = node.textValue();
-        if (text == null) {
-            return null;
-        }
-
-        final SchemaLocation schemaLocation = SchemaLocation.of(node.textValue());
-
-        return validationContext.getJsonSchemaFactory().getSchema(schemaLocation, validationContext.getConfig());
+        this.errorMessageType = errorMessageType;
+        this.evaluationPath = evaluationPath;
+        this.evaluationParentSchema = evaluationParentSchema;
     }
 
     protected static boolean equals(double n1, double n2) {
@@ -143,129 +129,6 @@ public abstract class BaseJsonValidator extends ValidationMessageHandler impleme
             builder.append(")");
             logger.debug(builder.toString());
         }
-    }
-
-    /**
-     * Checks based on the current {@link DiscriminatorContext} whether the provided {@link JsonSchema} a match against
-     * the current discriminator.
-     *
-     * @param currentDiscriminatorContext the currently active {@link DiscriminatorContext}
-     * @param discriminator               the discriminator to use for the check
-     * @param discriminatorPropertyValue  the value of the <code>discriminator/propertyName</code> field
-     * @param jsonSchema                  the {@link JsonSchema} to check
-     */
-    protected static void checkDiscriminatorMatch(final DiscriminatorContext currentDiscriminatorContext,
-                                                  final ObjectNode discriminator,
-                                                  final String discriminatorPropertyValue,
-                                                  final JsonSchema jsonSchema) {
-        if (discriminatorPropertyValue == null) {
-            currentDiscriminatorContext.markIgnore();
-            return;
-        }
-
-        final JsonNode discriminatorMapping = discriminator.get("mapping");
-        if (null == discriminatorMapping) {
-            checkForImplicitDiscriminatorMappingMatch(currentDiscriminatorContext,
-                    discriminatorPropertyValue,
-                    jsonSchema);
-        } else {
-            checkForExplicitDiscriminatorMappingMatch(currentDiscriminatorContext,
-                    discriminatorPropertyValue,
-                    discriminatorMapping,
-                    jsonSchema);
-            if (!currentDiscriminatorContext.isDiscriminatorMatchFound()
-                    && noExplicitDiscriminatorKeyOverride(discriminatorMapping, jsonSchema)) {
-                checkForImplicitDiscriminatorMappingMatch(currentDiscriminatorContext,
-                        discriminatorPropertyValue,
-                        jsonSchema);
-            }
-        }
-    }
-
-    /**
-     * Rolls up all nested and compatible discriminators to the root discriminator of the type. Detects attempts to redefine
-     * the <code>propertyName</code> or mappings.
-     *
-     * @param currentDiscriminatorContext the currently active {@link DiscriminatorContext}
-     * @param discriminator               the discriminator to use for the check
-     * @param schema                      the value of the <code>discriminator/propertyName</code> field
-     * @param instanceLocation                          the logging prefix
-     */
-    protected static void registerAndMergeDiscriminator(final DiscriminatorContext currentDiscriminatorContext,
-                                                        final ObjectNode discriminator,
-                                                        final JsonSchema schema,
-                                                        final JsonNodePath instanceLocation) {
-        final JsonNode discriminatorOnSchema = schema.schemaNode.get("discriminator");
-        if (null != discriminatorOnSchema && null != currentDiscriminatorContext
-                .getDiscriminatorForPath(schema.schemaLocation)) {
-            // this is where A -> B -> C inheritance exists, A has the root discriminator and B adds to the mapping
-            final JsonNode propertyName = discriminatorOnSchema.get("propertyName");
-            if (null != propertyName) {
-                throw new JsonSchemaException(instanceLocation + " schema " + schema + " attempts redefining the discriminator property");
-            }
-            final ObjectNode mappingOnContextDiscriminator = (ObjectNode) discriminator.get("mapping");
-            final ObjectNode mappingOnCurrentSchemaDiscriminator = (ObjectNode) discriminatorOnSchema.get("mapping");
-            if (null == mappingOnContextDiscriminator && null != mappingOnCurrentSchemaDiscriminator) {
-                // here we have a mapping on a nested discriminator and none on the root discriminator, so we can simply
-                // make it the root's
-                discriminator.set("mapping", discriminatorOnSchema);
-            } else if (null != mappingOnContextDiscriminator && null != mappingOnCurrentSchemaDiscriminator) {
-                // here we have to merge. The spec doesn't specify anything on this, but here we don't accept redefinition of
-                // mappings that already exist
-                final Iterator<Map.Entry<String, JsonNode>> fieldsToAdd = mappingOnCurrentSchemaDiscriminator.fields();
-                while (fieldsToAdd.hasNext()) {
-                    final Map.Entry<String, JsonNode> fieldToAdd = fieldsToAdd.next();
-                    final String mappingKeyToAdd = fieldToAdd.getKey();
-                    final JsonNode mappingValueToAdd = fieldToAdd.getValue();
-
-                    final JsonNode currentMappingValue = mappingOnContextDiscriminator.get(mappingKeyToAdd);
-                    if (null != currentMappingValue && currentMappingValue != mappingValueToAdd) {
-                        throw new JsonSchemaException(instanceLocation + "discriminator mapping redefinition from " + mappingKeyToAdd
-                                + "/" + currentMappingValue + " to " + mappingValueToAdd);
-                    } else if (null == currentMappingValue) {
-                        mappingOnContextDiscriminator.set(mappingKeyToAdd, mappingValueToAdd);
-                    }
-                }
-            }
-        }
-        currentDiscriminatorContext.registerDiscriminator(schema.schemaLocation, discriminator);
-    }
-
-    private static void checkForImplicitDiscriminatorMappingMatch(final DiscriminatorContext currentDiscriminatorContext,
-                                                                  final String discriminatorPropertyValue,
-                                                                  final JsonSchema schema) {
-        if (schema.schemaLocation.getFragment().getName(-1).equals(discriminatorPropertyValue)) {
-            currentDiscriminatorContext.markMatch();
-        }
-    }
-
-    private static void checkForExplicitDiscriminatorMappingMatch(final DiscriminatorContext currentDiscriminatorContext,
-                                                                  final String discriminatorPropertyValue,
-                                                                  final JsonNode discriminatorMapping,
-                                                                  final JsonSchema schema) {
-        final Iterator<Map.Entry<String, JsonNode>> explicitMappings = discriminatorMapping.fields();
-        while (explicitMappings.hasNext()) {
-            final Map.Entry<String, JsonNode> candidateExplicitMapping = explicitMappings.next();
-            if (candidateExplicitMapping.getKey().equals(discriminatorPropertyValue)
-                    && ("#" + schema.schemaLocation.getFragment().toString())
-                            .equals(candidateExplicitMapping.getValue().asText())) {
-                currentDiscriminatorContext.markMatch();
-                break;
-            }
-        }
-    }
-
-    private static boolean noExplicitDiscriminatorKeyOverride(final JsonNode discriminatorMapping,
-                                                              final JsonSchema parentSchema) {
-        final Iterator<Map.Entry<String, JsonNode>> explicitMappings = discriminatorMapping.fields();
-        while (explicitMappings.hasNext()) {
-            final Map.Entry<String, JsonNode> candidateExplicitMapping = explicitMappings.next();
-            if (candidateExplicitMapping.getValue().asText()
-                    .equals(parentSchema.schemaLocation.getFragment().toString())) {
-                return false;
-            }
-        }
-        return true;
     }
 
     @Override
@@ -313,14 +176,6 @@ public abstract class BaseJsonValidator extends ValidationMessageHandler impleme
         return getParentSchema();
     }
 
-    protected JsonSchema fetchSubSchemaNode(ValidationContext validationContext) {
-        return this.suppressSubSchemaRetrieval ? null : obtainSubSchemaNode(this.schemaNode, validationContext);
-    }
-
-    public Set<ValidationMessage> validate(ExecutionContext executionContext, JsonNode node) {
-        return validate(executionContext, node, node, atRoot());
-    }
-
     protected String getNodeFieldType() {
         JsonNode typeField = this.getParentSchema().getSchemaNode().get("type");
         if (typeField != null) {
@@ -335,42 +190,7 @@ public abstract class BaseJsonValidator extends ValidationMessageHandler impleme
         }
     }
 
-    public static class JsonNodePathLegacy {
-        private static final JsonNodePath INSTANCE = new JsonNodePath(PathType.LEGACY);
-        public static JsonNodePath getInstance() {
-            return INSTANCE;
-        }
-    }
 
-    public static class JsonNodePathJsonPointer {
-        private static final JsonNodePath INSTANCE = new JsonNodePath(PathType.JSON_POINTER);
-        public static JsonNodePath getInstance() {
-            return INSTANCE;
-        }
-    }
-
-    public static class JsonNodePathJsonPath {
-        private static final JsonNodePath INSTANCE = new JsonNodePath(PathType.JSON_PATH);
-        public static JsonNodePath getInstance() {
-            return INSTANCE;
-        }
-    }
-
-    /**
-     * Get the root path.
-     *
-     * @return The path.
-     */
-    protected JsonNodePath atRoot() {
-        if (this.validationContext.getConfig().getPathType().equals(PathType.JSON_POINTER)) {
-            return JsonNodePathJsonPointer.getInstance();
-        } else if (this.validationContext.getConfig().getPathType().equals(PathType.LEGACY)) {
-            return JsonNodePathLegacy.getInstance();
-        } else if (this.validationContext.getConfig().getPathType().equals(PathType.JSON_PATH)) {
-            return JsonNodePathJsonPath.getInstance();
-        }
-        return new JsonNodePath(this.validationContext.getConfig().getPathType());
-    }
 
     @Override
     public String toString() {
@@ -408,9 +228,14 @@ public abstract class BaseJsonValidator extends ValidationMessageHandler impleme
         return false;
     }
 
-    @Override
     protected MessageSourceValidationMessage.Builder message() {
-        return super.message().schemaNode(this.schemaNode);
+        return MessageSourceValidationMessage.builder(this.validationContext.getConfig().getMessageSource(), this.errorMessage, (message, failFast) -> {
+            if (failFast) {
+                throw new FailFastAssertionException(message);
+            }
+        }).code(this.errorMessageType.getErrorCode()).schemaNode(this.schemaNode).schemaLocation(this.schemaLocation)
+                .evaluationPath(this.evaluationPath).type(this.keyword != null ? this.keyword.getValue() : null)
+                .messageKey(this.errorMessageType.getErrorCodeValue());
     }
 
     /**
