+++ src/main/java/com/networknt/schema/DiscriminatorValidator.java
@@ -81,4 +81,127 @@ public class DiscriminatorValidator extends BaseJsonValidator {
     public Map<String, String> getMapping() {
         return mapping;
     }
+
+    /**
+     * Checks based on the current {@link DiscriminatorContext} whether the provided {@link JsonSchema} a match against
+     * the current discriminator.
+     *
+     * @param currentDiscriminatorContext the currently active {@link DiscriminatorContext}
+     * @param discriminator               the discriminator to use for the check
+     * @param discriminatorPropertyValue  the value of the <code>discriminator/propertyName</code> field
+     * @param jsonSchema                  the {@link JsonSchema} to check
+     */
+    public static void checkDiscriminatorMatch(final DiscriminatorContext currentDiscriminatorContext,
+                                                  final ObjectNode discriminator,
+                                                  final String discriminatorPropertyValue,
+                                                  final JsonSchema jsonSchema) {
+        if (discriminatorPropertyValue == null) {
+            currentDiscriminatorContext.markIgnore();
+            return;
+        }
+
+        final JsonNode discriminatorMapping = discriminator.get("mapping");
+        if (null == discriminatorMapping) {
+            checkForImplicitDiscriminatorMappingMatch(currentDiscriminatorContext,
+                    discriminatorPropertyValue,
+                    jsonSchema);
+        } else {
+            checkForExplicitDiscriminatorMappingMatch(currentDiscriminatorContext,
+                    discriminatorPropertyValue,
+                    discriminatorMapping,
+                    jsonSchema);
+            if (!currentDiscriminatorContext.isDiscriminatorMatchFound()
+                    && noExplicitDiscriminatorKeyOverride(discriminatorMapping, jsonSchema)) {
+                checkForImplicitDiscriminatorMappingMatch(currentDiscriminatorContext,
+                        discriminatorPropertyValue,
+                        jsonSchema);
+            }
+        }
+    }
+
+    /**
+     * Rolls up all nested and compatible discriminators to the root discriminator of the type. Detects attempts to redefine
+     * the <code>propertyName</code> or mappings.
+     *
+     * @param currentDiscriminatorContext the currently active {@link DiscriminatorContext}
+     * @param discriminator               the discriminator to use for the check
+     * @param schema                      the value of the <code>discriminator/propertyName</code> field
+     * @param instanceLocation                          the logging prefix
+     */
+    public static void registerAndMergeDiscriminator(final DiscriminatorContext currentDiscriminatorContext,
+                                                        final ObjectNode discriminator,
+                                                        final JsonSchema schema,
+                                                        final JsonNodePath instanceLocation) {
+        final JsonNode discriminatorOnSchema = schema.schemaNode.get("discriminator");
+        if (null != discriminatorOnSchema && null != currentDiscriminatorContext
+                .getDiscriminatorForPath(schema.schemaLocation)) {
+            // this is where A -> B -> C inheritance exists, A has the root discriminator and B adds to the mapping
+            final JsonNode propertyName = discriminatorOnSchema.get("propertyName");
+            if (null != propertyName) {
+                throw new JsonSchemaException(instanceLocation + " schema " + schema + " attempts redefining the discriminator property");
+            }
+            final ObjectNode mappingOnContextDiscriminator = (ObjectNode) discriminator.get("mapping");
+            final ObjectNode mappingOnCurrentSchemaDiscriminator = (ObjectNode) discriminatorOnSchema.get("mapping");
+            if (null == mappingOnContextDiscriminator && null != mappingOnCurrentSchemaDiscriminator) {
+                // here we have a mapping on a nested discriminator and none on the root discriminator, so we can simply
+                // make it the root's
+                discriminator.set("mapping", discriminatorOnSchema);
+            } else if (null != mappingOnContextDiscriminator && null != mappingOnCurrentSchemaDiscriminator) {
+                // here we have to merge. The spec doesn't specify anything on this, but here we don't accept redefinition of
+                // mappings that already exist
+                final Iterator<Map.Entry<String, JsonNode>> fieldsToAdd = mappingOnCurrentSchemaDiscriminator.fields();
+                while (fieldsToAdd.hasNext()) {
+                    final Map.Entry<String, JsonNode> fieldToAdd = fieldsToAdd.next();
+                    final String mappingKeyToAdd = fieldToAdd.getKey();
+                    final JsonNode mappingValueToAdd = fieldToAdd.getValue();
+
+                    final JsonNode currentMappingValue = mappingOnContextDiscriminator.get(mappingKeyToAdd);
+                    if (null != currentMappingValue && currentMappingValue != mappingValueToAdd) {
+                        throw new JsonSchemaException(instanceLocation + "discriminator mapping redefinition from " + mappingKeyToAdd
+                                + "/" + currentMappingValue + " to " + mappingValueToAdd);
+                    } else if (null == currentMappingValue) {
+                        mappingOnContextDiscriminator.set(mappingKeyToAdd, mappingValueToAdd);
+                    }
+                }
+            }
+        }
+        currentDiscriminatorContext.registerDiscriminator(schema.schemaLocation, discriminator);
+    }
+
+    private static void checkForImplicitDiscriminatorMappingMatch(final DiscriminatorContext currentDiscriminatorContext,
+                                                                  final String discriminatorPropertyValue,
+                                                                  final JsonSchema schema) {
+        if (schema.schemaLocation.getFragment().getName(-1).equals(discriminatorPropertyValue)) {
+            currentDiscriminatorContext.markMatch();
+        }
+    }
+
+    private static void checkForExplicitDiscriminatorMappingMatch(final DiscriminatorContext currentDiscriminatorContext,
+                                                                  final String discriminatorPropertyValue,
+                                                                  final JsonNode discriminatorMapping,
+                                                                  final JsonSchema schema) {
+        final Iterator<Map.Entry<String, JsonNode>> explicitMappings = discriminatorMapping.fields();
+        while (explicitMappings.hasNext()) {
+            final Map.Entry<String, JsonNode> candidateExplicitMapping = explicitMappings.next();
+            if (candidateExplicitMapping.getKey().equals(discriminatorPropertyValue)
+                    && ("#" + schema.schemaLocation.getFragment().toString())
+                            .equals(candidateExplicitMapping.getValue().asText())) {
+                currentDiscriminatorContext.markMatch();
+                break;
+            }
+        }
+    }
+
+    private static boolean noExplicitDiscriminatorKeyOverride(final JsonNode discriminatorMapping,
+                                                              final JsonSchema parentSchema) {
+        final Iterator<Map.Entry<String, JsonNode>> explicitMappings = discriminatorMapping.fields();
+        while (explicitMappings.hasNext()) {
+            final Map.Entry<String, JsonNode> candidateExplicitMapping = explicitMappings.next();
+            if (candidateExplicitMapping.getValue().asText()
+                    .equals(parentSchema.schemaLocation.getFragment().toString())) {
+                return false;
+            }
+        }
+        return true;
+    }    
 }
